Replace binary parser with the already compiled wheel
