+++ .github/workflows/pytest.yml
@@ -16,7 +16,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          pip install wheel setuptools pip pybind11 --upgrade
+          pip install pip --upgrade
           pip install -r ./requirements/dev.txt
           pip install pytest
       - name: Build tests
